A better fix for the json marshalling/unmarshalling of vcs-config.
The issue I ran into earlier was simple, I needed VcsConfig to be
a *json.RawMessage instead of a json.RawMessage. But I did not want
to just fix that because it would have sent the vcs-config as JSON
to the client exposing any credentials that are stored there. I use
a trick here where I create a special type like json.RawMessage that
will preserve bytes in Unmarshal, but just emit {} in marshal.