Remove lowercasing of repo names
The lowercasing of repo names is entirely unnecessary, and only serves
to complicate things for anyone trying to use the rest api. By keeping
the repo names intact, there is no longer a need to first call the repos
endpoint in order to lookup an actual name, if the name is the only
piece of information that is needed.