searcher: modify concurrent routine
Change the concurrent searchers init routine to simplify the
implementation and use only one common channel to return successfully
created searchers as well as errors. This way we also get rid of the go
routine collecting the results on the channel by making the main routine
collect the results by blocking the execution and listening on the
channel.