Remove old React version and import React, ReactDOM and createReactClass directly from the entry-points
In this commit (af4d9a89a73218a25834c4bdab994c5c1481fb74) React and ReactDOM where added as dependencies of the project
But it seems that they are not used because no entry-point import them. The React version that is used is the version 0.12.2 that is coming from 2014 and it is loaded as a separate script directly from the HTML templates.
The purpose of this commit is to import React, ReactDOM, and createReactClass from the entry-points and get rid of the old react version.
In this way, the project can be escalated using the new features of React