Back to short peek buf, with double read encoding fallback
The big peek buf was needed to ensure that a non-utf8 rune at the end of
the file does get the attention and be tried with the fallback encoding.

A more robust approach is to just try the reading as-is, validating the
encoding, and try the fallback encoding if this fails.

Also add a test case.