Fix the linter, update golang versions. (#422)
* golangci to v3 per https://stackoverflow.com/questions/71758856/github-action-for-golangci-lint-fails-with-cant-load-fmt

* add with latest

* add setup-go and specific version

* update version

* Use only one actions/checkout, previous version of go linter

The linter (`golangci`) and the action that runs the linter (`golangci/golangci-lint-action`) are separate. The linter seems to be running but finding new errors, so this change might be enough to temporarily squash those.

* Drop go 1.12 and 1.13,  add 1.17 and 1.18

There's a new golang version every 6 months. The latest two golang versions are supported officially. I'm keeping 1.14 around for now since it was previously the latest supported version.

* Only run linter on 1.17

The 1.18 linter has import issues that we'll investigate later.

* update version

* nolint

* nolint

* nolint

* nolint

* update lint to go 1.18

Co-authored-by: David Schott <dschott@etsy.com>